refactor: update import paths and parser file name
